+++ packages/service-library/src/servicelib/aiohttp/rest_middlewares.py
@@ -16,7 +16,7 @@ from models_library.error_codes import create_error_code
 from models_library.errors_classes import OsparcErrorMixin
 from models_library.utils.json_serialization import json_dumps
 
-from ..logging_utils import create_troubleshotting_log_message, get_log_record_extra
+from ..logging_errors import create_troubleshotting_log_message, get_log_record_extra
 from ..mimetype_constants import MIMETYPE_APPLICATION_JSON
 from ..utils import is_production_environ
 from .rest_models import ErrorItemType, ErrorType, LogMessageType
