+++ packages/service-library/tests/test_logging_utils.py
@@ -6,14 +6,10 @@ from typing import Any
 
 import pytest
 from faker import Faker
-from models_library.error_codes import create_error_code
-from models_library.errors_classes import OsparcErrorMixin
 from servicelib.logging_utils import (
     LogExtra,
     LogLevelInt,
     LogMessageStr,
-    create_troubleshotting_log_message,
-    get_log_record_extra,
     guess_message_log_level,
     log_context,
     log_decorator,
@@ -381,42 +377,3 @@ def test_set_parent_module_log_level_(caplog: pytest.LogCaptureFixture):
 
     assert "parent warning" in caplog.text
     assert "child warning" in caplog.text
-
-
-def test_create_troubleshotting_log_message(caplog: pytest.LogCaptureFixture):
-    class MyError(OsparcErrorMixin, RuntimeError):
-        msg_template = "My error {user_id}"
-
-    with pytest.raises(MyError) as exc_info:
-        raise MyError(user_id=123, product_name="foo")
-
-    exc = exc_info.value
-    error_code = create_error_code(exc)
-    log_msg = create_troubleshotting_log_message(
-        f"Nice message to user [{error_code}]",
-        exc,
-        error_code=error_code,
-        error_context=exc.error_context(),
-        tip="This is a test error",
-    )
-
-    with caplog.at_level(logging.WARNING):
-        root_logger = logging.getLogger()
-        root_logger.exception(
-            log_msg, extra=get_log_record_extra(error_code=error_code)
-        )
-
-        # ERROR    root:test_logging_utils.py:417 Nice message to user [OEC:126055703573984].
-        # {
-        # "exception_details": "My error 123",
-        # "error_code": "OEC:126055703573984",
-        # "context": {
-        #     "user_id": 123,
-        #     "product_name": "foo"
-        # },
-        # "tip": "This is a test error"
-        # }
-
-        assert error_code in caplog.text
-        assert "user_id" in caplog.text
-        assert "product_name" in caplog.text
